refactor: extract workspace clone command
